test: split recorder e2e specs
